refactor: move SeeAllCard into ui/cards
